Seed Database

+++ CBU2SINIFPROJE.Entities/Concrete/Employee.cs
@@ -6,7 +6,7 @@ namespace CBU2SINIFPROJE.Entities.Concrete
 {
     public class Employee : Person, IEmployee
     {
-        public decimal Salary { get; set; }
+        public int Salary { get; set; }
         public List<Vacation> Vacations { get; set; }
     }
 }
